+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.0.2] - 2025-04-18
+## [1.0.2] - 2025-04-24
+
+### Added
+
+- (Apr 24) Updated content
+
+  - Achievements: Added new HackerRank certifications
+  - Footer: Included HackerRank profile link
 
 ### Changed